css/css-flexbox/flexbox_first-letter-align-items.html

<!DOCTYPE html>
<html lang="en">
<head>
  <title>Flexbox | First Letter</title>
  <meta name="assert" content="Flexbox align-items does not affect the pseudo element ::first-letter">
  <link rel="help" href="https://drafts.csswg.org/css-pseudo-4/#selectordef-first-letter">
  <link rel="author" title="Clayton Salem" href="claytonjsalem@gmail.com">
  <style>
    #first-letter-test {
      display: flex;
    }
    #first-letter-test p::first-letter {
      align-items: baseline;
      font-size: 25px;
    }
  </style>
</head>
<body>
  <div id="first-letter-test">
    <p>Lorem ipsum dolor sit amet.</p>
  </div>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script>
    test(function() {
      const firstLetterStyle = getComputedStyle(document.querySelector('p'), ':first-letter')
      // Double check we're looking at the correct pseudo element.
      assert_equals(firstLetterStyle.fontSize, '25px');
      // Check we're not setting the flexbox align-items value.
      assert_not_equals(firstLetterStyle.alignItems, 'baseline');
    })
  </script>
</body>
</html>
